Update GNodePaths.java
moved method call paths() outside of main method, and wrote code to test the method.  the result is a print out of all the paths, where the names of GNodes in a path are printed on a line from root to leaf.  each path is printed on a new line.